Time: 64 ms (12.22%), Space: 8.9 MB (68.39%) - LeetHub

+++ 423-reconstruct-original-digits-from-english/423-reconstruct-original-digits-from-english.cpp
@@ -19,11 +19,11 @@ public:
         }
         
         for (size_t i = 0; i <= 9; i++) {
-            int idx = hint[i] - 'a';
+            size_t idx = hint[i] - 'a';
             int count = cnt[idx];
             
             for (size_t j = 0; j < word[i].length(); j++) {
-                cnt[word[i][j]-'a']-=count;
+                cnt[word[i][j] - 'a'] -= count;
             }
             while (count) {    
                 ans+=to_string(num[i]);
